refactor: 어노테이션 경로 재정의
