+++ src/main/java/com/tasksprints/auction/bid/application/resolver/BidValidationResolver.java
@@ -1,0 +1,4 @@
+package com.tasksprints.auction.bid.application.resolver;
+
+public class BidValidationResolver {
+}
