+++ src/main/java/com/tasksprints/auction/chat/application/resolver/WhisperValidationResolver.java
@@ -1,7 +1,7 @@
 package com.tasksprints.auction.chat.application.resolver;
 
 import com.tasksprints.auction.chat.application.service.ChatService;
-import com.tasksprints.auction.chat.domain.annotation.WhisperValidation;
+import com.tasksprints.auction.chat.application.annotation.WhisperValidation;
 import com.tasksprints.auction.chat.domain.dto.WhisperDto;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
